+++ angular-material-app/src/app/components/header/header.component.ts
@@ -14,7 +14,7 @@ import { MatDialog } from "@angular/material/dialog";
 import { SearchDialogComponent } from "../../partials/search-dialog/search-dialog.component";
 import { SettingsDialogComponent } from "../../partials/settings-dialog/settings-dialog.component";
 import { FollowDialogComponent } from "../../partials/follow-dialog/follow-dialog.component";
-import { BlogSocialIconsComponent } from "../blog-social-icons/blog-social-icons.component";
+import { BlogSocialIconsComponent } from "../../partials/blog-social-icons/blog-social-icons.component";
 
 import { MatSlideToggleModule } from "@angular/material/slide-toggle";
 import { MatIconModule } from "@angular/material/icon";
